Finished QA bot development.
Completed the QA bot development, training and prediction phases. It is trained for multinomial responses but practically, it will only function as a binary since the outputs are only yes and no. REMEMBER: the questions are limited by the vocabulary. So, it is not going to run properly if you go beyond the limited library contained in the dataset.